refactor: 회원가입 API의 URI를 /auth로 변경 - 게이트웨이에서 로그인과 함께 분리시키기 위해 설정
