+++ src/main/java/com/baedalping/delivery/domain/auth/dto/UserCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.baedalping.delivery.domain.user.dto.request;
+package com.baedalping.delivery.domain.auth.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
